refactor: move things to rsc-utils
