+++ src/middleware/lib/common.ts
@@ -1,10 +1,6 @@
 import type { Config, Middleware } from "../../config.js";
 
 export type Shared = {
-  generatePrefetchCode?: (
-    entryItems: Iterable<readonly [rscId: string, props: unknown]>,
-    moduleIds: Iterable<string>
-  ) => string;
   devScriptToInject?: (path: string) => Promise<string>;
   prdScriptToInject?: (path: string) => Promise<string>;
 };
