+++ src/middleware/rewriteRsc.ts
@@ -1,80 +1,8 @@
-import { Buffer } from "node:buffer";
-import { Transform } from "node:stream";
-
 import type { MiddlewareCreator } from "./lib/common.js";
 
 // This convension is just one idea.
 
-// HACK It's not ideal to export a function from middleware.
-export const generatePrefetchCode = (
-  entryItemsIterable: Iterable<readonly [rscId: string, props: unknown]>,
-  moduleIds: Iterable<string>
-) => {
-  const entryItems = Array.from(entryItemsIterable);
-  let code = "";
-  if (entryItems.length) {
-    const rscIds = [...new Set(entryItems.map(([rscId]) => rscId))];
-    code += `
-globalThis.__WAKUWORK_PREFETCHED__ = {
-${rscIds
-  .map((rscId) => {
-    const value =
-      "{" +
-      entryItems
-        .flatMap(([id, props]) => {
-          if (id !== rscId) return [];
-          // FIXME we blindly expect JSON.stringify usage is deterministic
-          const serializedProps = JSON.stringify(props);
-          const searchParams = new URLSearchParams();
-          searchParams.set("props", serializedProps);
-          return [
-            `'${serializedProps}': fetch('/RSC/${rscId}/${searchParams}')`,
-          ];
-        })
-        .join(",") +
-      "}";
-    return `  '${rscId}': ${value}`;
-  })
-  .join(",\n")}
-};`;
-  }
-  for (const moduleId of moduleIds) {
-    code += `
-import('${moduleId}');`;
-  }
-  return code;
-};
-
-// HACK Patching stream is very fragile.
-// HACK No reason to have this function in this file
-export const transformRsfId = (
-  prefixToRemove: string,
-  convert = (id: string) => id
-) =>
-  new Transform({
-    transform(chunk, encoding, callback) {
-      if (encoding !== ("buffer" as any)) {
-        throw new Error("Unknown encoding");
-      }
-      const data = chunk.toString();
-      const lines = data.split("\n");
-      let changed = false;
-      for (let i = 0; i < lines.length; ++i) {
-        const match = lines[i].match(
-          new RegExp(`^([0-9]+):{"id":"${prefixToRemove}(.*?)"(.*)$`)
-        );
-        if (match) {
-          lines[i] = `${match[1]}:{"id":"${convert(match[2])}"${match[3]}`;
-          changed = true;
-        }
-      }
-      callback(null, changed ? Buffer.from(lines.join("\n")) : chunk);
-    },
-  });
-
-const rewriteRsc: MiddlewareCreator = (_config, shared) => {
-  shared.generatePrefetchCode = generatePrefetchCode;
-
+const rewriteRsc: MiddlewareCreator = () => {
   return async (req, _res, next) => {
     const url = new URL(req.url || "", "http://" + req.headers.host);
     if (url.pathname.startsWith("/RSC/")) {
